Added an Android-based RC control
Added a new node, rc_socket_2_roscopter which listes to socket connections and receives RC commands from there and then sends it to roscopter. Also copied an Android app (quadcontrol.pak) to the rc_control_android_app folder. This can be used as RC control. This .apk is frm ghowen.me, thanks to him for the app.